Fixing the sharing disclaimer text. Moving firebase config to the environment files.
