+++ src/environments/environment.prod.ts
@@ -19,3 +19,15 @@ export const environment = {
   authClientID: '419125973937-kl2cru5pu2vfugne7lr1hosgseh4lo1s.apps.googleusercontent.com',
   authScope: 'https://www.googleapis.com/auth/bigquery'
 };
+
+// Your web app's Firebase configuration
+var firebaseConfig = {
+  apiKey: "AIzaSyDS8k-x7L9vZ_mvvdyTzwQ1LNXsYLNnhOM",
+  authDomain: "bigquerygeoviz.firebaseapp.com",
+  databaseURL: "https://bigquerygeoviz.firebaseio.com",
+  projectId: "bigquerygeoviz",
+  storageBucket: "bigquerygeoviz.appspot.com",
+  messagingSenderId: "419125973937",
+  appId: "1:419125973937:web:eba1c63d64b58be3ec2390",
+  measurementId: "G-FNH2K1BP5G"
+};
